+++ x/inflation/types/msgs.go
@@ -4,27 +4,30 @@ import (
 	"fmt"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/x/auth/migrations/legacytx"
 )
 
 // ensure Msg interface compliance at compile time
 var (
-	_ sdk.Msg = &MsgEditInflationParams{}
-	_ sdk.Msg = &MsgToggleInflation{}
+	_ legacytx.LegacyMsg = &MsgEditInflationParams{}
+	_ legacytx.LegacyMsg = &MsgToggleInflation{}
+	_ legacytx.LegacyMsg = &MsgBurn{}
 )
 
 // oracle message types
 const (
 	TypeMsgEditInflationParams = "edit_inflation_params"
 	TypeMsgToggleInflation     = "toggle_inflation"
+	TypeMsgBurn                = "msg_burn"
 )
 
-// Route implements sdk.Msg
+// Route implements legacytx.LegacyMsg
 func (msg MsgEditInflationParams) Route() string { return RouterKey }
 
-// Type implements sdk.Msg
+// Type implements legacytx.LegacyMsg
 func (msg MsgEditInflationParams) Type() string { return TypeMsgEditInflationParams }
 
-// GetSignBytes implements sdk.Msg
+// GetSignBytes implements legacytx.LegacyMsg
 func (msg MsgEditInflationParams) GetSignBytes() []byte {
 	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(&msg))
 }
@@ -75,13 +78,13 @@ func (m MsgEditInflationParams) ValidateBasic() error {
 
 // -------------------------------------------------
 // MsgToggleInflation
-// Route implements sdk.Msg
+// Route implements legacytx.LegacyMsg
 func (msg MsgToggleInflation) Route() string { return RouterKey }
 
-// Type implements sdk.Msg
+// Type implements legacytx.LegacyMsg
 func (msg MsgToggleInflation) Type() string { return TypeMsgToggleInflation }
 
-// GetSignBytes implements sdk.Msg
+// GetSignBytes implements legacytx.LegacyMsg
 func (msg MsgToggleInflation) GetSignBytes() []byte {
 	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(&msg))
 }
@@ -100,5 +103,44 @@ func (m MsgToggleInflation) ValidateBasic() error {
 	if _, err := sdk.AccAddressFromBech32(m.Sender); err != nil {
 		return err
 	}
+	return nil
+}
+
+// -------------------------------------------------
+// MsgBurn
+// Route implements legacytx.LegacyMsg
+func (msg MsgBurn) Route() string { return RouterKey }
+
+// Type implements legacytx.LegacyMsg
+func (msg MsgBurn) Type() string { return TypeMsgBurn }
+
+// GetSignBytes implements legacytx.LegacyMsg
+func (msg MsgBurn) GetSignBytes() []byte {
+	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(&msg))
+}
+
+// GetSigners implements legacytx.LegacyMsg
+func (msg MsgBurn) GetSigners() []sdk.AccAddress {
+	feeder, err := sdk.AccAddressFromBech32(msg.Sender)
+	if err != nil {
+		panic(err)
+	}
+
+	return []sdk.AccAddress{feeder}
+}
+
+func (m MsgBurn) ValidateBasic() error {
+	if _, err := sdk.AccAddressFromBech32(m.Sender); err != nil {
+		return err
+	}
+
+	if err := m.Coin.Validate(); err != nil {
+		return err
+	}
+
+	if m.Coin.Amount.IsZero() {
+		return fmt.Errorf("coin amount should not be zero")
+	}
+
 	return nil
 }